Module ue4pak::archive

Raw FArchive tools

Structs

ArchiveLen
A write archive wrapper that count written bytes

ArchiveReader
A read archive wrapper for io::Read

ArchiveWriter
A write archive wrapper for io::Write

Traits

Archivable
A data structure that can be archived (encoded/decoded)

ArchivableWith

Archive
An archive reader or writer trait There is a single trait in order to simplify Archivable impls.