Module ue4pak::constants

Constants

COMPRESSION_METHOD_NAME_LEN

Length of a compression format name

COMPRESS_BIAS_MEMORY

Prefer compression that compresses smaller (ONLY VALID FOR COMPRESSION)

COMPRESS_BIAS_SPEED

Prefer compression that compresses faster (ONLY VALID FOR COMPRESSION)

COMPRESS_CUSTOM

Compress with user defined callbacks - DEPRECATED, USE FNAME

COMPRESS_DEPRECATED_FORMAT_FLAGS_MASK

Joint of the previous ones to determine if old flags are being used

COMPRESS_GZIP

Compress with GZIP - DEPRECATED, USE FNAME

COMPRESS_NONE

No compression

COMPRESS_NO_FLAGS

No flags specified /

COMPRESS_OPTIONS_FLAGS_MASK

Set of flags that are options are still allowed

COMPRESS_SOURCE_IS_PADDED

Is the source buffer padded out (ONLY VALID FOR UNCOMPRESS)

COMPRESS_ZLIB

Compress with ZLIB - DEPRECATED, USE FNAME

MAX_CHUNK_DATA_SIZE

Size of cached data.

MAX_NUM_COMPRESSION_METHODS

Number of allowed different methods

PAK_FILE_MAGIC

Magic number to use in header