Struct ue4pak::PakIndexV2

pub struct PakIndexV2 {
    pub mount_point: String,
    pub path_hash_seed: u64,
    pub has_path_hash_index: bool,
    pub path_hash_index_hash: [u8; 20],
    pub has_full_directory_index: bool,
    pub full_directory_index_hash: [u8; 20],
    /* private fields */
}

Fields
mount_point: String

Mount point

path_hash_seed: u64

The seed passed to the hash function for hashing filenames in this pak. Differs per pack so that the same filename in different paks has different hashes

has_path_hash_index: boolpath_hash_index_hash: [u8; 20]has_full_directory_index: boolfull_directory_index_hash: [u8; 20]
Implementations

impl PakIndexV2

pub fn clear(&mut self)

pub fn add(
&mut self,
name: String,
entry: PakEntry,
version: PakVersion
) -> Result<PakEntryLocation>

pub fn hashed_entries(&self) -> impl Iterator<Item = (u64, &PakEntry)> + '_

pub fn entries(&self) -> impl Iterator<Item = &PakEntry>

pub fn pruned_entries(
&self
) -> impl Iterator<Item = (&str, &str, PakEntryLocation)>

pub fn full_entries(
&self
) -> impl Iterator<Item = (&str, &str, PakEntryLocation)>

pub fn ser<A: Archive>(&mut self, ar: &mut A, version: PakVersion) -> Result<()>

pub fn de<A: Archive + Seek>(
&mut self,
ar: &mut A,
version: PakVersion
) -> Result<()>

pub fn ser_de<A, F>(
&mut self,
ar: &mut A,
version: PakVersion,
seek: F
) -> Result<()>where
A: Archive,
F: FnMut(&mut A, u64, u64, [u8; 20], &'static str) -> Result<()>,